src/RAG/requirements.txt
streamlit
sentence-transformers
faiss-cpu
requests
pdfplumber
python-docx
beautifulsoup4
html2text
transformers
nbformat
nltk